src/client/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
cramjam==2.7.0
cycler==0.11.0
fastparquet==2023.7.0
filelock==3.12.0
fonttools==4.42.1
fsspec==2023.6.0
grpcio==1.54.2
grpcio-tools==1.54.2
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.2
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
zipp==3.16.2
